// src/app/components/tokenPrice/types.ts
export interface TokenPrice {
  price: number;
  priceChain: number;
  price5m: number;
  priceChain5m: number;
  variation5m: number;
  price1h: number;
  priceChain1h: number;
  variation1h: number;
  price6h: number;
  priceChain6h: number;
  variation6h: number;
  price24h: number;
  priceChain24h: number;
  variation24h: number;
}